requirements.txt
yfinance
pandas
openpyxl
xlrd
requests
google-auth
google-auth-oauthlib
google-api-python-client
